+++ Database/DatabaseSetup.cs
@@ -5,21 +5,24 @@ namespace LabManager.Database;
 class DatabaseSetup
 {
 
-    public DatabaseSetup()
+    private readonly DatabaseConfig _databaseConfig;
+
+    public DatabaseSetup(DatabaseConfig DatabaseConfig)
     {
+        _databaseConfig = DatabaseConfig;
         CreateComputerTable();
         CreateLabTable();
     }
     private void CreateComputerTable()
     {
-        var connection = new SqliteConnection("Data Source=database.db");
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
         connection.Open();
         var command = connection.CreateCommand();
         command.CommandText = @"
         CREATE TABLE IF NOT EXISTS Computers(
-            id int not null primary key,
-            ram varchar(100) not null,
-            processor varchar(100) not null
+        id int not null primary key,
+        ram varchar(100) not null,
+        processor varchar(100) not null
         );
         ";
         command.ExecuteNonQuery();
@@ -32,10 +35,10 @@ class DatabaseSetup
         command = connection.CreateCommand();
         command.CommandText = @"
         CREATE TABLE IF NOT EXISTS Lab(
-            id int not null primary key,
-            number int not null,
-            name varchar(100) not null,
-            block varchar(50) not null
+        id int not null primary key,
+        number int not null,
+        name varchar(100) not null,
+        block varchar(50) not null
         );
         ";
         command.ExecuteNonQuery();
