+++ LabManager.csproj
