+++ Program.cs
@@ -2,9 +2,11 @@ using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
 
-var databaseSetup = new DatabaseSetup();
+var databaseConfig = new DatabaseConfig();
 
-var computerRepository = new ComputerRepository();
+var databaseSetup = new DatabaseSetup(databaseConfig);
+
+var computerRepository = new ComputerRepository(databaseConfig);
 
 // Routing
 var modelName = args[0];
@@ -13,31 +15,28 @@ var modelAction = args[1];
 
 if (modelName == "Computer")
 {
-    if (modelAction == "List")
+    if (modelAction == "List") // roteamento
     {
-           System.Console.WriteLine("Computer List");
-           foreach (var computer in computerRepository.GetAll())
-           {
-               System.Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
-           }
+        System.Console.WriteLine("Computer List");
+        foreach (var computer in computerRepository.GetAll())
+        {
+            System.Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+        }
     }
     if (modelAction == "New")
     {
-            int id = Convert.ToInt32(args[2]);
-            string ram = args[3];
-            string processor = args[4];
-
-            var connection = new SqliteConnection("Data Source=database.db");
-            connection.Open();
-
-            var command = connection.CreateCommand();
-            command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
-            command.Parameters.AddWithValue("$id", id);
-            command.Parameters.AddWithValue("$ram", ram);
-            command.Parameters.AddWithValue("$processor", processor);
-            command.ExecuteNonQuery();
-
-            connection.Close();
+        int id = Convert.ToInt32(args[2]);
+        string ram = args[3];
+        string processor = args[4];
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
+        connection.Open();
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", id);
+        command.Parameters.AddWithValue("$ram", ram);
+        command.Parameters.AddWithValue("$processor", processor);
+        command.ExecuteNonQuery();
+        connection.Close();
     }
 
     if(modelAction == "Delete")
